Bingo Seed
Added the bingo seed as an input box.
This makes it more obvious that it is copyable, and allows a user to enter in a new seed at the same time. I am currently looking into ways to parse the entered string into new variables to be read back. Will use an if (ex. if NewSeed[1] = 2 then gui,add,picture, blahblah picture2.png) which will be similar to ButtonNewGame:

Also standardizing version numbers. This is version 1 (fully functional with the same base code) .7 (feature addition 7) .5 (includes work for next version) .10.4.18 (the date)