1.8a.10.5.2
>Cleaned code (literally added more {} )
>Performed some additional testing for v1.8
>titles will now be version number

Thoughts: it appears submitting the seed will not allow a new loop to review the variable name with a %A_Index% in it. Other alternative to a loop is a rats nest if statement (re-doing version 1 specifically for this button).

Another thought: A change to the version numbering is necessary for multiple commits in a day.  Add a number for the day revision. The numbers look bizarre now, so this will be 1.8a.10.5.2 (a for in-progress b for ready. No more year, then rev on that day.)

regarding the earlier submission, co-author is Joe