1.8b.10.5.3
Seeds are now fully functional. Seeds can be changed multiple times per run. A new ahk file is located in resources. This will unfortunately need to be compiled separately to use without install. file name at line #1094 will also need to be changed to myseed.exe at that point.

The new ahk file in resources is a copy of the initial commit source code with some name changes. It is run by using #Include %NameOfFile%. It is slower than the NewGame: loop and will continue to be a problem as new tiles are made for the cards.